Fix temp file check for Hanamogeta (and potentially other games)

See merge request i-wanna-play-online/modder!7

+++ gml/worldGameEnd.gml
@@ -1,6 +1,6 @@
 /// ONLINE
 #if TEMPFILE
-	if(!file_exists("temp") && !file_exists("temp.dat")){
+	if(!file_exists("temp") && !file_exists(working_directory+"\save\temp") && !file_exists("temp.dat")){
 		if(file_exists("tempOnline")){
 			file_delete("tempOnline");
 		}
